gcp_functions/requirements .txt
pandas
requests
cloud-sql-python-connector[pg8000]
google-cloud-secret-manager
psycopg2-binary
pyarrow